UID2 Shared Module: Moved over shared uid2 code from module to library to remove lint error (#13025)
* moved shared code into library

* adjusted imports to be from libraries

* fixed importws for ajax and utils

* renamed folder name
